// src/enums/Sound.Type.enum.ts
export enum SoundType {
    SOUND_AHHH = 'sound-ahhh',
    SOUND_BOCCHI = 'sound-bocchi',
    SOUND_CAU_MET_LAM_HA = 'sound-cau-met-lam-ha',
    SOUND_GAMBUSTA = 'sound-gambusta',
    SOUND_HAYAY = 'sound-hayay',
    SOUND_OH_MY_GOD = 'sound-oh-my-god',
    SOUND_SHINDERU = 'sound-shinderu',
    SOUND_SUGOI_SUGOI = 'sound-sugoi-sugoi',
    SOUND_WAOO_CONGRATULATIONS = 'sound-waoo-congratulations',
}